+++ src/test/groovy/com/epam/reportportal/spock/fail/ThreeStepsOneFailedTest.java
@@ -30,7 +30,7 @@ import com.epam.ta.reportportal.ws.model.FinishTestItemRQ;
 import com.epam.ta.reportportal.ws.model.StartTestItemRQ;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.runner.Result;
+import org.junit.platform.launcher.listeners.TestExecutionSummary;
 import org.mockito.ArgumentCaptor;
 
 import java.util.List;
@@ -47,6 +47,7 @@ import static org.mockito.Mockito.*;
 
 public class ThreeStepsOneFailedTest {
 
+	private final String launchId = CommonUtils.namedId("launch_");
 	private final String classId = CommonUtils.namedId("class_");
 	private final List<String> methodIds = Stream.generate(() -> CommonUtils.namedId("method_")).limit(3).collect(Collectors.toList());
 
@@ -54,16 +55,16 @@ public class ThreeStepsOneFailedTest {
 
 	@BeforeEach
 	public void setupMock() {
-		TestUtils.mockLaunch(client, null, classId, methodIds);
+		TestUtils.mockLaunch(client, launchId, classId, methodIds);
 		TestUtils.mockBatchLogging(client);
 		TestExtension.listener = new ReportPortalSpockListener(ReportPortal.create(client, standardParameters(), testExecutor()));
 	}
 
 	@Test
 	public void verify_one_failed_unrolled_step_reporting() {
-		Result result = runClasses(HelloSpockSpecUnrollFailed.class);
+		TestExecutionSummary result = runClasses(HelloSpockSpecUnrollFailed.class);
 
-		assertThat(result.getFailureCount(), equalTo(1));
+		assertThat(result.getTotalFailureCount(), equalTo(1L));
 
 		verify(client).startLaunch(any());
 		ArgumentCaptor<StartTestItemRQ> startCaptor = ArgumentCaptor.forClass(StartTestItemRQ.class);
@@ -86,6 +87,7 @@ public class ThreeStepsOneFailedTest {
 				.collect(Collectors.toList());
 		assertThat(finishItemStatuses, containsInAnyOrder(ItemStatus.PASSED.name(), ItemStatus.PASSED.name(), ItemStatus.FAILED.name()));
 		verify(client).finishTestItem(eq(classId), any());
+		verify(client).finishLaunch(eq(launchId), any());
 
 		//noinspection unchecked
 		verify(client, atLeast(1)).log(any(List.class));