+++ src/test/groovy/com/epam/reportportal/spock/fixtures/TestCleanupFixtureFailureIntegrity.java
@@ -29,7 +29,7 @@ import com.epam.ta.reportportal.ws.model.FinishTestItemRQ;
 import com.epam.ta.reportportal.ws.model.StartTestItemRQ;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.runner.Result;
+import org.junit.platform.launcher.listeners.TestExecutionSummary;
 import org.mockito.ArgumentCaptor;
 
 import java.util.List;
@@ -44,6 +44,7 @@ import static org.mockito.ArgumentMatchers.same;
 import static org.mockito.Mockito.*;
 
 public class TestCleanupFixtureFailureIntegrity {
+	private final String launchId = CommonUtils.namedId("launch_");
 	private final String classId = CommonUtils.namedId("class_");
 	private final List<String> methodIds = Stream.generate(() -> CommonUtils.namedId("method_")).limit(2).collect(Collectors.toList());
 
@@ -51,16 +52,16 @@ public class TestCleanupFixtureFailureIntegrity {
 
 	@BeforeEach
 	public void setupMock() {
-		TestUtils.mockLaunch(client, null, classId, methodIds);
+		TestUtils.mockLaunch(client, launchId, classId, methodIds);
 		TestUtils.mockBatchLogging(client);
 		TestExtension.listener = new ReportPortalSpockListener(ReportPortal.create(client, standardParameters(), testExecutor()));
 	}
 
 	@Test
 	public void verify_setup_fixture_failure_correct_reporting() {
-		Result result = runClasses(CleanupFixtureFailed.class);
+		TestExecutionSummary result = runClasses(CleanupFixtureFailed.class);
 
-		assertThat(result.getFailureCount(), equalTo(1));
+		assertThat(result.getTotalFailureCount(), equalTo(1L));
 
 		verify(client).startLaunch(any());
 		verify(client).startTestItem(any(StartTestItemRQ.class));
@@ -84,6 +85,7 @@ public class TestCleanupFixtureFailureIntegrity {
 		});
 
 		verify(client).finishTestItem(eq(classId), any());
+		verify(client).finishLaunch(eq(launchId), any());
 		//noinspection unchecked
 		verify(client).log(any(List.class));
 		verifyNoMoreInteractions(client);