+++ src/test/groovy/com/epam/reportportal/spock/fixtures/TestCleanupSpecFixtureFailureIntegrity.java
@@ -21,7 +21,6 @@ import com.epam.reportportal.listeners.ItemType;
 import com.epam.reportportal.service.ReportPortal;
 import com.epam.reportportal.service.ReportPortalClient;
 import com.epam.reportportal.spock.ReportPortalSpockListener;
-import com.epam.reportportal.spock.features.fixtures.CleanupFixtureFailed;
 import com.epam.reportportal.spock.features.fixtures.CleanupSpecFixtureFailed;
 import com.epam.reportportal.spock.utils.TestExtension;
 import com.epam.reportportal.spock.utils.TestUtils;
@@ -30,7 +29,7 @@ import com.epam.ta.reportportal.ws.model.FinishTestItemRQ;
 import com.epam.ta.reportportal.ws.model.StartTestItemRQ;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.runner.Result;
+import org.junit.platform.launcher.listeners.TestExecutionSummary;
 import org.mockito.ArgumentCaptor;
 
 import java.util.List;
@@ -45,6 +44,7 @@ import static org.mockito.ArgumentMatchers.same;
 import static org.mockito.Mockito.*;
 
 public class TestCleanupSpecFixtureFailureIntegrity {
+	private final String launchId = CommonUtils.namedId("launch_");
 	private final String classId = CommonUtils.namedId("class_");
 	private final List<String> methodIds = Stream.generate(() -> CommonUtils.namedId("method_")).limit(2).collect(Collectors.toList());
 
@@ -52,16 +52,16 @@ public class TestCleanupSpecFixtureFailureIntegrity {
 
 	@BeforeEach
 	public void setupMock() {
-		TestUtils.mockLaunch(client, null, classId, methodIds);
+		TestUtils.mockLaunch(client, launchId, classId, methodIds);
 		TestUtils.mockBatchLogging(client);
 		TestExtension.listener = new ReportPortalSpockListener(ReportPortal.create(client, standardParameters(), testExecutor()));
 	}
 
 	@Test
 	public void verify_setup_fixture_failure_correct_reporting() {
-		Result result = runClasses(CleanupSpecFixtureFailed.class);
+		TestExecutionSummary result = runClasses(CleanupSpecFixtureFailed.class);
 
-		assertThat(result.getFailureCount(), equalTo(1));
+		assertThat(result.getTotalFailureCount(), equalTo(1L));
 
 		verify(client).startLaunch(any());
 		verify(client).startTestItem(any(StartTestItemRQ.class));
@@ -86,6 +86,8 @@ public class TestCleanupSpecFixtureFailureIntegrity {
 		ArgumentCaptor<FinishTestItemRQ> finishSpecCaptor = ArgumentCaptor.forClass(FinishTestItemRQ.class);
 		verify(client).finishTestItem(eq(classId), finishSpecCaptor.capture());
 		assertThat(finishSpecCaptor.getValue().getStatus(), equalTo(ItemStatus.FAILED.name()));
+
+		verify(client).finishLaunch(eq(launchId), any());
 
 		//noinspection unchecked
 		verify(client).log(any(List.class));