+++ src/test/groovy/com/epam/reportportal/spock/nestedsteps/ThreePassedNestedStepsTest.java
@@ -30,7 +30,7 @@ import com.epam.ta.reportportal.ws.model.StartTestItemRQ;
 import org.apache.commons.lang3.tuple.Pair;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.runner.Result;
+import org.junit.platform.launcher.listeners.TestExecutionSummary;
 import org.mockito.ArgumentCaptor;
 
 import java.util.List;
@@ -65,9 +65,9 @@ public class ThreePassedNestedStepsTest {
 
 	@Test
 	public void verify_passed_nested_step_reporting() {
-		Result result = runClasses(HelloSpockSpec.class);
+		TestExecutionSummary result = runClasses(HelloSpockSpec.class);
 
-		assertThat(result.getFailureCount(), equalTo(0));
+		assertThat(result.getTotalFailureCount(), equalTo(0L));
 
 		ArgumentCaptor<StartTestItemRQ> startCaptor = ArgumentCaptor.forClass(StartTestItemRQ.class);
 		verify(client).startTestItem(any(StartTestItemRQ.class));