+++ src/test/groovy/com/epam/reportportal/spock/params/NullParamsNameTest.java
@@ -27,7 +27,7 @@ import com.epam.ta.reportportal.ws.model.StartTestItemRQ;
 import org.apache.commons.lang3.tuple.Pair;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.runner.Result;
+import org.junit.platform.launcher.listeners.TestExecutionSummary;
 import org.mockito.ArgumentCaptor;
 
 import java.util.List;
@@ -62,8 +62,8 @@ public class NullParamsNameTest {
 
 	@Test
 	public void verify_null_parameter_step_name() {
-		Result result = runClasses(NullParamSpec.class);
-		assertThat(result.getFailureCount(), equalTo(0));
+		TestExecutionSummary result = runClasses(NullParamSpec.class);
+		assertThat(result.getTotalFailureCount(), equalTo(0L));
 
 		ArgumentCaptor<StartTestItemRQ> startCaptor = ArgumentCaptor.forClass(StartTestItemRQ.class);
 		verify(client, timeout(2000).times(2)).startTestItem(same(methodId), startCaptor.capture());