+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Client version updated on [5.1.16](https://github.com/reportportal/client-java/releases/tag/5.1.16), by @HardNorth
+- `NodeInfoUtils` class refactored to not use synchronized Map, by @HardNorth 
 
 ## [5.1.0]
 ### Added